fix(api): address comments
Ref. metriport/metriport-internal#2526
