--- packages/core/src/command/ai-brief/ai-brief-shared.ts
@@ -1,38 +1,0 @@
-import { Binary, Bundle } from "@medplum/fhirtypes";
-import { Buffer } from "buffer";
-import dayjs from "dayjs";
-import { uuidv7 } from "../../util/uuid-v7";
-
-export async function generateAiBriefFhirResource(
-  content: string | undefined
-): Promise<Binary | undefined> {
-  if (!content) return undefined;
-
-  const encodedContent = Buffer.from(content).toString("base64");
-
-  return {
-    resourceType: "Binary",
-    id: uuidv7(),
-    meta: {
-      versionId: "1",
-      lastUpdated: dayjs().toISOString(),
-      source: "metriport:ai-generated-brief",
-    },
-    contentType: "text/plain",
-    data: encodedContent,
-  };
-}
-
-export function getAiBriefContentFromBundle(bundle: Bundle): string | undefined {
-  const binaryResourceEntry = bundle.entry?.find(
-    entry => entry.resource?.resourceType === "Binary"
-  );
-
-  if (!binaryResourceEntry) return undefined;
-
-  const binaryResource = binaryResourceEntry.resource as Binary;
-
-  if (!binaryResource.data) return undefined;
-
-  return Buffer.from(binaryResource.data, "base64").toString("utf-8");
-}
