+++ packages/core/src/command/consolidated/consolidated-create.ts
@@ -12,8 +12,8 @@ import { executeAsynchronously, out } from "../../util";
 import { Config } from "../../util/config";
 import { getConsolidatedLocation, getConsolidatedSourceLocation } from "./consolidated-shared";
 import { isAiBriefFeatureFlagEnabledForCx } from "../../external/aws/app-config";
-import { summarizeFilteredBundleWithAI } from "../../command/ai-brief/ai-brief-create";
-import { generateAiBriefFhirResource } from "../../command/ai-brief/ai-brief-shared";
+import { summarizeFilteredBundleWithAI } from "../ai-brief/create";
+import { generateAiBriefFhirResource } from "../ai-brief/shared";
 
 const s3Utils = new S3Utils(Config.getAWSRegion());
 
@@ -48,9 +48,10 @@ export async function createConsolidatedFromConversions({
   const fhirPatient = patientToFhir(patient);
   const patientEntry = buildBundleEntry(fhirPatient);
 
-  const [conversions, docRefs] = await Promise.all([
+  const [conversions, docRefs, isAiBriefFeatureFlagEnabled] = await Promise.all([
     getConversions({ cxId, patient, sourceBucketName }),
     getDocumentReferences({ cxId, patientId }),
+    isAiBriefFeatureFlagEnabledForCx(cxId),
   ]);
   log(`Got ${conversions.length} resources from conversions`);
 
@@ -65,12 +66,11 @@ export async function createConsolidatedFromConversions({
   const deduped = deduplicate({ cxId, patientId, bundle: withDups });
   log(`...done, from ${withDups.entry?.length} to ${deduped.entry?.length} resources`);
 
-  const isAiBriefFeatureFlagEnabled = await isAiBriefFeatureFlagEnabledForCx(cxId);
   log(`isAiBriefFeatureFlagEnabled: ${isAiBriefFeatureFlagEnabled}`);
 
   if (isAiBriefFeatureFlagEnabled) {
     const aiBriefContent = await summarizeFilteredBundleWithAI(withDups, cxId, patientId);
-    const aiBriefFhirResource = await generateAiBriefFhirResource(aiBriefContent);
+    const aiBriefFhirResource = generateAiBriefFhirResource(aiBriefContent);
     if (aiBriefFhirResource) {
       deduped.entry?.push(buildBundleEntry(aiBriefFhirResource));
     }
