+++ packages/core/src/external/aws/lambda-logic/bundle-to-html-shared.ts
@@ -1,6 +1,6 @@
 import { DiagnosticReport } from "@medplum/fhirtypes";
 import dayjs from "dayjs";
-import { Brief } from "../../../command/ai-brief/ai-brief-create";
+import { Brief } from "../../../command/ai-brief/create";
 
 export const ISO_DATE = "YYYY-MM-DD";
 export const MISSING_DATE_KEY = "N/A";
