+++ packages/core/src/external/fhir/shared/bundle.ts
@@ -33,6 +33,7 @@ import {
   ServiceRequest,
 } from "@medplum/fhirtypes";
 import { filterTruthy } from "@metriport/shared/common/filter-map";
+import { isBinary } from ".";
 import { SearchSetBundle } from "@metriport/shared/medical";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
@@ -254,7 +255,7 @@ export function extractFhirTypesFromBundle(bundle: Bundle): ExtractedFhirTypes {
       const resource = entry.resource;
       if (resource?.resourceType === "Patient") {
         patient = resource as Patient;
-      } else if (resource?.resourceType === "Binary") {
+      } else if (isBinary(resource)) {
         binaries.push(resource as Binary);
       } else if (resource?.resourceType === "DocumentReference") {
         documentReferences.push(resource as DocumentReference);
